refactor(omegaform): extract public type definitions
